feat: comprehensive codebase cleanup and restore animated CTA buttons
## Code Cleanup
- Remove unused components: BlogGrid, CategoryFilter, FeaturedPost, Footer, Header, ScrollProgress
- Delete duplicate UI components in src/components/ui/ directory
- Remove unused pages directory (src/pages/)
- Clean up unnecessary UI components: squares-background, ripple-grid, grid-pattern, etc.
- Remove unused imports and fix type definitions

## CTA Button Restoration
- Restore original animated HoverBorderGradient component with rotating border effect
- Fix framer-motion animations and proper gradient calculations
- Ensure seamless background blending with theme colors
- Maintain consistent styling across hero section and footer

## Build Fixes
- Fix CustomSeparator import errors by replacing with inline CSS
- Update import paths after directory cleanup
- Fix theme provider imports and type definitions
- Resolve NavLink type reference errors

## Result
- Build compiles successfully with no errors
- All original functionality preserved
- Animated gradient border CTA buttons working perfectly
- Cleaner, more maintainable codebase structure
