--- src/components/CategoryFilter.tsx
@@ -1,38 +1,0 @@
-import React, { useState } from 'react';
-import { cn } from '../../lib/utils';
-
-interface CategoryFilterProps {
-  categories: string[];
-  activeCategory: string;
-  onCategoryChange: (category: string) => void;
-}
-
-const CategoryFilter: React.FC<CategoryFilterProps> = ({
-  categories,
-  activeCategory,
-  onCategoryChange
-}) => {
-  const allCategories = ['all', ...categories];
-
-  return (
-    <div className="flex flex-wrap gap-2 mb-8">
-      {allCategories.map((category) => (
-        <button
-          key={category}
-          onClick={() => onCategoryChange(category)}
-          className={cn(
-            "px-4 py-2 rounded-full text-sm font-medium transition-colors duration-200",
-            "border border-gray-200 hover:border-gray-300",
-            activeCategory === category
-              ? "bg-blue-600 text-white border-blue-600"
-              : "bg-white text-gray-700 hover:bg-gray-50"
-          )}
-        >
-          {category === 'all' ? 'All Posts' : category.charAt(0).toUpperCase() + category.slice(1)}
-        </button>
-      ))}
-    </div>
-  );
-};
-
-export default CategoryFilter;
